+++ internal/controller/suite_test.go
@@ -19,19 +19,14 @@ package controller
 import (
 	"context"
 	"path/filepath"
-	"strings"
 	"testing"
 
 	"cloud.google.com/go/pubsub"
-	"cloud.google.com/go/pubsub/apiv1/pubsubpb"
 	"cloud.google.com/go/pubsub/pstest"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
+	"github.com/quipper/google-cloud-pubsub-operator/internal/pubsubtest"
 	"google.golang.org/api/option"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/credentials/insecure"
-	"google.golang.org/grpc/status"
 	"k8s.io/client-go/kubernetes/scheme"
 	"k8s.io/client-go/rest"
 	ctrl "sigs.k8s.io/controller-runtime"
@@ -95,20 +90,14 @@ var _ = BeforeSuite(func() {
 	Expect(err).ToNot(HaveOccurred())
 
 	psServer = pstest.NewServer(
-		createTopicErrorInjectionReactor(),
+		pubsubtest.CreateTopicErrorInjectionReactor(),
 	)
 	DeferCleanup(func() {
 		Expect(psServer.Close()).Should(Succeed())
 	})
 
 	newClient := func(ctx context.Context, projectID string, opts ...option.ClientOption) (c *pubsub.Client, err error) {
-		return pubsub.NewClient(ctx, projectID,
-			append(opts,
-				option.WithEndpoint(psServer.Addr),
-				option.WithoutAuthentication(),
-				option.WithGRPCDialOption(grpc.WithTransportCredentials(insecure.NewCredentials())),
-			)...,
-		)
+		return pubsubtest.NewClient(ctx, projectID, psServer)
 	}
 
 	err = (&TopicReconciler{
@@ -131,24 +120,3 @@ var _ = BeforeSuite(func() {
 		Expect(err).ToNot(HaveOccurred(), "failed to run manager")
 	}()
 })
-
-func createTopicErrorInjectionReactor() pstest.ServerReactorOption {
-	return pstest.ServerReactorOption{
-		FuncName: "CreateTopic",
-		Reactor: errorInjectionReactorFunc(func(req interface{}) (bool, interface{}, error) {
-			topic, ok := req.(*pubsubpb.Topic)
-			if ok {
-				if strings.HasPrefix(topic.Name, "projects/error-injected-") {
-					return true, nil, status.Errorf(codes.InvalidArgument, "error injected")
-				}
-			}
-			return false, nil, nil
-		}),
-	}
-}
-
-type errorInjectionReactorFunc func(req interface{}) (bool, interface{}, error)
-
-func (r errorInjectionReactorFunc) React(req interface{}) (bool, interface{}, error) {
-	return r(req)
-}
