+++ internal/controller/topic_controller_test.go
@@ -4,13 +4,10 @@ import (
 	"context"
 	"time"
 
-	"cloud.google.com/go/pubsub"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	googlecloudpubsuboperatorv1 "github.com/quipper/google-cloud-pubsub-operator/api/v1"
-	"google.golang.org/api/option"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/credentials/insecure"
+	"github.com/quipper/google-cloud-pubsub-operator/internal/pubsubtest"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 	//+kubebuilder:scaffold:imports
@@ -20,11 +17,7 @@ var _ = Describe("Topic controller", func() {
 	Context("When creating a Topic resource", func() {
 		It("Should create a Pub/Sub Topic", func(ctx context.Context) {
 			const projectID = "topic-project-1"
-			psClient, err := pubsub.NewClient(ctx, projectID,
-				option.WithEndpoint(psServer.Addr),
-				option.WithoutAuthentication(),
-				option.WithGRPCDialOption(grpc.WithTransportCredentials(insecure.NewCredentials())),
-			)
+			psClient, err := pubsubtest.NewClient(ctx, projectID, psServer)
 			Expect(err).ShouldNot(HaveOccurred())
 
 			By("Creating a Topic")
